ml-service/requirements.txt
flask==3.0.3
tensorflow==2.17.0
pillow==10.4.0
numpy==2.1.1
gunicorn==22.0.0

